ARM: tegra: usb_phy: Set OC_CONTROLLER properly
The OC_CONTROLLER should be programmed first before any of the phy
registers are programmed. Updating the OC_CONTROLLER in the later
part of the phy_on is causing a glitch which results in the hsic
devices to be disconnected.

Bug 1212563

Change-Id: I5c6d1bd918c620301462c62cb569ef5a32c2713b
Signed-off-by: Suresh Mangipudi <smangipudi@nvidia.com>
Reviewed-on: http://git-master/r/189121
(cherry picked from commit e4a1ac3c9cd83dc904f34d842500340f5a5e7512)
Reviewed-on: http://git-master/r/190143
Reviewed-by: Mark Kuo <mkuo@nvidia.com>
Tested-by: Mark Kuo <mkuo@nvidia.com>
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>