ARM: configs: Add TRUSTED_FOUNDATIONS to molly_defconfig
Nvidia's Android makefile adds this on command line
based on build config option of SECURE_OS.  We
don't need such flexibility and will always
enable this so every build works without special
cmd line flags.  If not enabled and bootloader is
configured for SECURE_OS, boot is very sluggish and
we see errors about three of the four cores not
coming up.

Change-Id: If6a8dba25354b6358c05c7c4df9a7644900bb6bc
Signed-off-by: Mike J. Chen <mjchen@google.com>