arch: arm: mach-tegra: Add molly board files and adjust defconfig
The molly board files are copied from dalmore and only a few
small changes initially, like removing some peripherals
we won't need like touchscreens.  A lot more pruning is needed
to really make it a molly but this is a start.  The kernel
built by molly_defconfig works with the dalmore bootloader.

Change-Id: I9271092d47144367697677836ca081df3cebc373
Signed-off-by: Mike J. Chen <mjchen@google.com>