drivers: mbtc: Make athome remote work more reliably
There appears to be a problem with sending our cmd to
start le in the middle of a the startup sequence bluedroid
sends after reset.  Either that or the chip isn't ready
to accept such a command yet.  All we see is that the
chip doesn't respond to our le start command and we're
hung.

Adding a 1s delay appears to make it work.  Not yet tested
at extensively so not certain whether this is guaranteed.
Should investigate more with Marvell.

Fix a bug with athome_bt_pkt_to_user() calling wrong API.

Also, fix some apparent bug in athome_bt_filter_num_comp_pkt()
and it's adjustment of the rx_len.

Signed-off-by: Mike J. Chen <mjchen@google.com>

Change-Id: If6af71cfafcdaee1347e8b1b952f51ecb0edf447