mach-tegra: molly: regulator cleanup
Start removing regulators we don't have in Molly or the
version of Dalmore we're using.  The Dalmores we have
do not have the MAX77663 PMIC so remove all references
to it from.

The TPS65090 will also be removed later because Molly
doesn't have one, but Dalmore does and we need it for
things like SDMMC3 right now.

Also remove regulators we don't ever plan to use.

Rename the wlan regulator to remove bcm reference.

Change-Id: I703827c31bf87fbc6a9ce3770067ad6f1c9fb7b8
Signed-off-by: Mike J. Chen <mjchen@google.com>