arm: boot: dts: Add dts file for molly
It's not really used for much yet by the kernel since
CONFIG_PINCTRL_TEGRA114 isn't even enabled (matches
Dalmore kernel config) but I tried to make it correct
in case we do want to switch to it in the future.

We should build it though, because we have a DTB
partition and need to put something into it (at least
the compatible fields).

Change-Id: I96c697d3e33cba4df694d6d373ed8d25a31b5218
Signed-off-by: Mike J. Chen <mjchen@google.com>