usb: tegra: when DCD detect timeout, consider as unknown charger
On Molly, it uses same connector for AC power and USB debug cable.  The
VBUS on the cable is sourced from AC, instead of USB debug cable.
In case user only connects AC adapter, but leave USB debug cable
disconeected, the D+/D- is floating on the cable, so the DCD detection
would be timeout'd.  In this case, we should consider the charger type
as "unknown", instead of CDP or SDP, to avoid misconfiguring the UDC
controller when the USB cable is not connected.

Bug: 9817732

Signed-off-by: Haley Teng <hteng@nvidia.com>