AahBtRemote: Fix bugs with response buffer size not being set correctly
Some helper functions were passed a buffer and then passed
the size of the ptr to the buffer instead of the full size
of the response buffer athome_bt_simple_cmd().  This causes
a warning to get tripped.  Fix all those cases and rename
the parameter when used as a function argument so it's
less likely to happen in the future.

Change-Id: I126831af86211f4a46c76f3bac27e6866f1a07cb
Signed-off-by: Mike J. Chen <mjchen@google.com>