video: tegra: dc: Add some asserts to WAR code
Added an assert to the WAR for disp clock not giving the freq
asked for.  Make sure the dummy value is +2 of the requested
value and that it hasn't wrapped to become 0 or 1 or 2.  Since we
don't have low level details of the bug, it's unclear if this
is strictly needed.

Change-Id: I0f46ae9463aa13f07224c9190b554568b257fb04
Signed-off-by: Mike J. Chen <mjchen@google.com>