mach: tegra: molly: Pull down DDC i2c lines when HPD active
The Tegra definition of a low signal is below the value that
the level shifter considers to be a low (i.e. sometimes the
level shifter provides a low that is interpreted by the
Tegra as a high) so we have to add a pull-down on the Tegra
side to make the low "lower".  Turn off the pull-down when
no HPD since it draws power (level shifter has internal
pull-ups, so there's a path from 5V to ground through the
two pulls).

Change-Id: I0c054e4cb9fc954a1e1a3cae9d62a007d93e99b9
Signed-off-by: Mike J. Chen <mjchen@google.com>