Revert "ALSA: hda: check validity of speaker allocation field"
This reverts commit f7ba8410a3986afea0cce9a21672c74f5e9af221.

I don't believe it is right to convert a 0 value for spk_alloc
to 0xffff.  A DVI only monitor won't support audio and this
would make it say it does.  Also, if no HDMI cable is even plugged
in this would falsely report audio was supported.

Any bug in reading ELD should be fixed and not hacked like the
original CL was doing.

Signed-off-by: Mike J. Chen <mjchen@google.com>

Conflicts:
	sound/pci/hda/patch_hdmi.c

Change-Id: I9d03197ceafb37dc27a8960040e2432d4826907c