ALSA: hda: Introduce hdmi_audio switch
Fixes race condition of HDA not being ready for Android audio
system to use it because it's not had a chance to parse ELD
data yet.

There's a delay between when HDMI driver reads edid
and sets the hdmi switch until when the HDA driver gets notified
of the ELD data being ready and is able to parse it.  We need
a second switch to notify Android that audio is ready to use
in this case, which we clear when HDA notices the hotplug remove,
and which we set when HDA has parsed the ELD data for a hotplug insert.

Change-Id: Ifccdb20f7e689b8940f6bdd7e279a8bb7bb80fa0
Signed-off-by: Mike J. Chen <mjchen@google.com>