BemoteAudioDriver: prevent hang in del_timer_sync
A timer is used to decode SBC data from Bemote or to feed zeros
if Bemote overflows. The timer calls snd_pcm_period_elapsed(),
which on rare occasion call snd_pcm_trigger(),
which calls del_timer_sync(), which hangs if called
under interrupt.

The fix is to only call del_timer_sync() if we are
not under interrupt. If we are under interrupt
then the timer callback will not reschedule the timer
because the timer_enabled flag will be false.

Note that the timer does not have any resources that
need to be reclaimed. The del_timer_sync() function
just cancels any pending timers. If we don't
want to cancel a timer then we don't have to delete anything.

Bug: 11260437
Change-Id: I6488de9b7cd72dedd4c62e3b00b2d714d297ba35
Signed-off-by: Phil Burk <philburk@google.com>