usb: tegra_udc: clear work queues before suspend
It was possible (although rare under normal conditions) for irq_work to
be scheduled right before the device would go to suspend, and to only
run halfway before suspend happens. As a result, it would continue its
execution upon resume on a non-powered UDC, resulting in a system
freeze.

This patch prevents this to happen by doing the following:
- Make sure the dr_controller_*() functions are called while udc->lock is
  held, to avoid updating udc->stopped in the middle of an interrupt
- Do not schedule IRQ works if udc->stopped is true
- Ensure all scheduled UDC works are completed before suspending.

Bug: 10919955
Change-Id: Iceac24aaf29dd310cb73bc1f28c3e5d6b9248af5
Signed-off-by: Alexandre Courbot <acourbot@nvidia.com>