tegra: molly: lower EDP limits for CPU and Core
Due to the tight power budget on Molly, we end up regularly in an
over-current situation which triggers a SOC_THERM alarm and throttles
the CPU and GPU frequencies down. We want to avoid this for both
performance consistency and power safety reasons.

This patch lowers the EDP limits on CPU and Core to a point where OC
conditions almost to not appear.

Bug 15471845

Signed-off-by: Alexandre Courbot <acourbot@nvidia.com>
(cherry picked from commit ebe960f6b6a0323d0ae1070335d5fceb535c7f95)

Change-Id: I985f06df28ab3a8caca2388cf986f8505b607c98