mmc: tegra: hard code best tap value as 0 for sdmmc1
Nvidia ASIC has reviewed on Molly and confirmed that tap=0 should be the
best tap value for sdmmc1 with Marvell SD8797 wifi/bt chip.  In the
current tap auto-tuning algorithm, we also see 0 is selected as tap
value for sdmmc1 on Molly.

This commit prevents any unknown situation which might (we have not seen
this kind of case yet) cause tap is not selected as 0 for sdmmc1 in the
future.  This commit also adds a warning log for the case of the best
tap value is not calculated as 0 by auto-tuning algorithm for sdmmc1, so
we can review on that devices in case it will be needed.

Change-Id: I3bfbb683b636eb16b819af117cc92b747e7a5722
Signed-off-by: Haley Teng <hteng@nvidia.com>