AahBtRemote: Change LED debugging to disconnect only
As we get closer to ship, we need to let userland start to
control the LED.  So we only override the userland requested
LED value if we see a disconnect of the remote, and we
restore the userland value on connect.

Also fixed compile errors when the hack is disabled completely.

Change-Id: Ia16a47e0b7081ae7640f5b6d31b19a5bf0e10235
Signed-off-by: Mike J. Chen <mjchen@google.com>