AahBtRemote: Add analog tunings to dpad and handle repeat
We now handle key repeat rate based on a finger's distance from
the origin point of a swipe gesture. This allows for finer tuned
repeat rates based on user interaction, and replaces the 50ms
repeat that the Android framework was imposing on us. The main goal
of this change in the short term is to allow UX to tweak parameters
as they wish so we can move towards a more properly defined set of
tunings to implement.

Change-Id: Ic8bd59f18be9c59ea4b943a6ca885495f32f6b6e