[hdmi] Disable HDMI 5V while Molly is suspended
This change disables the HDMI 5V line while Molly is suspended.  The 5V line is
re-enabled when the DC enables HDMI.  This fixes the issue where "smart" HDMI
switches were getting confused because Molly leave the 5V on all the time.

Note that the 5V line is not renabled in molly_hdmi_hotplug_init() because
we do not want HPD to be active while Molly is suspended.

Signed-off-by:  Steve Paik <spaik@google.com>

Change-Id: I8108c3de5a03af137c4ea08d137a3c83f0195768