Android alarm:Eliminate the conflict with Linux alarmtimer
Android does not support powering up of phone through alarm clock. The
original code conflict with the Linux alarmtimer which was disabled.
Latest bluedroid would use the Linux alarmtimer (if exist) for better
handling. Eliminating the conflicts would make room for the future.
Hoping to keep the existing feature for later ports.